Вынес логику feedback-card в useFeedbackCard
