+++ src/entities/feedback/model/index.ts
@@ -1,0 +1,1 @@
+export * from './feedback-card-hook';
