+++ src/entities/feedback/ui/feedback-card.tsx
@@ -1,49 +1,20 @@
-import { ReactNode, useEffect, useState } from 'react';
+import { ReactNode } from 'react';
 import { motion } from 'motion/react';
+import { useFeedbackCard } from '../model';
 
-const FeedbackCard = ({
-	color,
-	children,
-	index,
-}: {
+type TFeedbackCard = {
 	color: string;
 	index: number;
 	children: ReactNode;
 	id?: number;
 	name?: string;
 	description?: string;
-}) => {
-	const [isHover, setIsHover] = useState(false);
-	const [isFull, setIsFull] = useState(false);
-	const [isMobile, setIsMobile] = useState(false);
-	const offset = isMobile ? 80 : 120;
+};
 
-	useEffect(() => {
-		const handleResize = () => {
-			setIsMobile(window.innerWidth <= 1024);
-		};
-		handleResize();
-		window.addEventListener('resize', handleResize);
-		return () => window.removeEventListener('resize', handleResize);
-	}, []);
-
-	const handleClick = () => {
-		setIsFull(prev => !prev);
-	};
-
-	const calculateLeft = () => {
-		if (isFull) {
-			return index * offset - offset;
-		}
-
-		if (isHover) {
-			return index * offset - offset;
-		} else {
-			return index * offset;
-		}
-	};
-
-	const left = calculateLeft();
+const FeedbackCard = ({ color, children, index }: TFeedbackCard) => {
+	const { handleClick, left, setIsHover, setIsFull, isFull } = useFeedbackCard({
+		index,
+	});
 
 	return (
 		<motion.div
