+++ src/widgets/promotions/ui/promotions.tsx
@@ -10,7 +10,7 @@ export const Promotions = () => {
 	return (
 		<SectionLayout id='promotions' className='mb-section-bottom lg:mb-24'>
 			<h2 className='font-intro text-blue-0 leading-[1] text-section-xl text-left mb-[75px] md:text-[40px]'>
-				АКТУАЛЬНЫЕ СКИДКИ <br className='md:hidden' /> И АКЦИИ
+				Актуальные скидки <br className='md:hidden' /> и акции
 			</h2>
 			<div className='mb-[125px] lg:mb-[90px] md:mb-[45px]'>
 				<CustomSwiper
